Add additional documentation and disable failure-no-check script until resolution of GH-18
