+++ src/util/nested-iterator-map.ts
@@ -1,4 +1,4 @@
-import { NestedIterable } from "..";
+import { NestedIterable } from "./nested-iterable";
 
 export function * nestedIterableMap<T>(iterator: NestedIterable<T>, mapping: (t: NestedIterable<T>, previous: T[]) => NestedIterable<T>, previous: T[] = []): NestedIterable<T> {
     for(const [t, nested] of mapping(iterator, previous)) {
